Changed elapsed time counting to float Moved Particle update logic to class

+++ operator.h
@@ -7,7 +7,7 @@ class Emitter;
 class Operator {
 public:
   virtual ~Operator() {};
-  virtual void apply(Particle *p, int elapsed) = 0;
+  virtual void apply(Particle *p, float elapsed) = 0;
 };
 
 #endif // OPERATOR_H
